Applied styling fix to Dropdown component instead
